test: add tests for recreation-resource service type filters
